SubscriptionManagementREDFISH.py: Rename File
- Renamed to reflect that it will now be able to manage subscriptions and SSE streaming
- Rename in separate commit so that changes are clear.
